[model] Use qf_model_t internally for IQM
Many things are broken (iqm models don't renderer visibly, vulkan can
crash with a validation error (something about indices)), but alias
models still work and IQM models load without any apparent memory
corruption (there was a lot during development).

+++ include/QF/GL/qf_alias.h
@@ -33,9 +33,9 @@
 #include "QF/GL/types.h"
 
 typedef struct aliasvrt_s {
-	GLshort     st[2];
-	GLshort     normal[3];
-	GLushort    vertex[3];
+	float       st[2];
+	float       normal[3];
+	float       vertex[3];
 } aliasvrt_t;
 
 struct entity_s;
